SF [ 1860228 ] LINUX: Even More GCC Warnings Fixes (thanks monkeyman)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@11127 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/GUIDialogSettings.cpp
@@ -139,7 +139,7 @@ void CGUIDialogSettings::SetupPage()
 
 void CGUIDialogSettings::EnableSettings(unsigned int id, bool enabled)
 {
-  unsigned int index = (unsigned int) -1;
+  int index = -1;
   for (unsigned int i = 0; i < m_settings.size(); i++)
   {
     if (m_settings[i].id != id)
